pyproject.toml
[project]
name = "fantasybaby-tennis-cut"
version = "0.1.0"
description = "AI-assisted tennis rally video cutter for removing dead time and keeping smooth points."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.9.0",
    "numpy>=1.24.0",
    "PyYAML>=6.0.0",
    "tqdm>=4.66.0",
]

[project.scripts]
tennis-cut = "fantasybaby_tennis_cut.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["fantasybaby_tennis_cut*"]
